Move utility folder to com.uwflow.flow_android.util
